yuml, unofficial command line client for yUML.me
This is an unofficial command line client for the yUML web app. Use it
to draw class diagrams, activity diagrams and usecase diagrams from
the command line, integrate yUML into your documentation workflow or
what have you.

yuml (the tool) supports all the styles and formats that yUML (the
service) does, so you can take your pick between scruffy (the default),
boring and plain for the styles and png, pdf, svg or jpg for the
format.

Have fun.